# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafe Shop API",
        "description": "API for managing menu, orders, and other cafe-related operations.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parse-news-sky": {
            "post": {
                "tags": [
                    "News Parsing"
                ],
                "summary": "Parse food news from news.sky.com and save it to the database.",
                "operationId": "9beac9efc871794883a82b1bde923418",
                "responses": {
                    "200": {
                        "description": "Successfully parsed and saved to the database.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to parse or save to the database.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Oops, there are temporary problems"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/parse-news-bbc": {
            "post": {
                "tags": [
                    "News Parsing"
                ],
                "summary": "Parse food news from bbc.co.uk and save it to the database.",
                "operationId": "10053472a2f145ff4aa180b9482face1",
                "responses": {
                    "200": {
                        "description": "Successfully parsed and saved to the database.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to parse or save to the database.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Oops, there are temporary problems"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payment": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment page url from Stripe",
                "operationId": "36ec1ffc6a03efe3a69c1159cc277325",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "type_order_id",
                                    "product",
                                    "success_url",
                                    "cancel_url"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type_order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product": {
                                        "required": [
                                            "id",
                                            "name",
                                            "price",
                                            "quantity"
                                        ],
                                        "properties": {
                                            "id": {
                                                "type": "array",
                                                "items": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "name": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "example": "Lattes"
                                                }
                                            },
                                            "price": {
                                                "type": "array",
                                                "items": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": 5
                                                }
                                            },
                                            "quantity": {
                                                "type": "array",
                                                "items": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "success_url": {
                                        "type": "string",
                                        "example": "http://localhost:3000/?success=true"
                                    },
                                    "cancel_url": {
                                        "type": "string",
                                        "example": "http://localhost:3000/?success=false"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stripe payment page URL",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "https://checkout.stripe.com/c/pay/efsdfdrgrdg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Oops, there are temporary problems",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Oops, there are temporary problems"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payment/webhook": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Handle Stripe webhook response",
                "operationId": "e3d80c0b0c142b0befc4293cca0bcc75",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Oops, there are temporary problems",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Oops, there are temporary problems"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}